+++ CHANGELOG.md
@@ -19,6 +19,13 @@ following sections:
 
 The versions follow [semantic versioning](https://semver.org).
 
+## Unreleased - YYYY-MM-DD
+
+### Changed
+
+- Made the workaround for `MachineReadableFormatError` introduced in 0.5.2 more
+  generic.
+
 ## 0.5.2 - 2019-10-27
 
 ### Added